es kompiliert wieder!
git-svn-id: https://psvn.ims.uni-hannover.de/svn/fucap/trunk@9 1233f08c-2788-c397-e086-bd64547a093a

+++ src/bm.cc
@@ -94,7 +94,7 @@ void EVAL_BM_ACTION_BASE::ac_final_adjust(T* y)const
   if (_bandwidth != NOT_INPUT && _bandwidth != 0.) {untested();
     assert(y->imag() == 0);
     double ratio = CKT_BASE::_sim->_freq / _bandwidth;
-    double coeff = y->real() / (1.+(ratio*ratio));
+    double coeff = double( y->real() / (1.+(ratio*ratio)) );
     *y = T(coeff, -coeff * ratio);
   }else{
   }
